Added comment

+++ src/main/java/CreditPaymentService.java
@@ -3,7 +3,7 @@ public class CreditPaymentService {
         double percentAtMonth = percent / 100 / 12;
         double koefOfPer = 1 + percentAtMonth;
         double koefOfPercent = Math.pow (koefOfPer, period);
-        int payment = (int) ((int) summa * (percentAtMonth + (percentAtMonth / (koefOfPercent - 1))));
+        int payment = (int) (summa * (percentAtMonth + (percentAtMonth / (koefOfPercent - 1))));
         return payment;
     }
 }
